+++ angular-workspace/projects/ni/nimble-angular/rich-text/editor/tests/nimble-rich-text-editor.directive.spec.ts
@@ -78,6 +78,16 @@ describe('Nimble Rich Text Editor', () => {
             expect(nativeElement.empty).toBeTrue();
         });
 
+        it('has valid configuration by default', () => {
+            expect(directive.checkValidity()).toBeTrue();
+            expect(nativeElement.checkValidity()).toBeTrue();
+        });
+
+        it('has all invalid configurations set to false by default', () => {
+            expect(directive.validity.invalidMentionConfiguration).toBeFalse();
+            expect(nativeElement.validity.duplicateMentionConfiguration).toBeFalse();
+        });
+
         it('has empty mentioned Hrefs array by default', () => {
             expect(directive.getMentionedHrefs()).toEqual([]);
             expect(nativeElement.getMentionedHrefs()).toEqual([]);