+++ packages/nimble-components/src/rich-text/base/types.ts
@@ -1,0 +1,6 @@
+import type { ValidityObject } from '../../utilities/models/validator';
+
+export interface RichTextValidity extends ValidityObject {
+    readonly invalidMentionConfiguration: boolean;
+    readonly duplicateMentionConfiguration: boolean;
+}